requirements.txt
langchain
OpenAI
tiktoken
python-dotenv